requirements.txt
# Data Quality Framework - Dependencies
# ======================================

# Database
mysql-connector-python>=8.0.33
PyMySQL>=1.1.0

# Configuration
PyYAML>=6.0.1

# Reporting
Jinja2>=3.1.2

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0

# Utilities
tabulate>=0.9.0
